+++ CHANGES.rst
@@ -1,3 +1,14 @@
+1.9.2dev
+--------
+
+- Modify tweak_standard for Mosfire/J2
+- Apply find_min_max when clipping the image for object finding
+- Mask bad detector regions for global sky flexure calculation
+- Detector structure correction included in flatfield calibration
+- Apply find_min_max when clipping the image for object finding
+- Mask bad detector regions for global sky flexure calculation
+- Refactor PypeIt input files
+- Added wavelength diagnostics to the spec2d output
 
 1.9.1dev
 --------